refactor: use openChildModally rather than presenting the analytics screen alone
